#673
- Error class added.

+++ olat2moodle.php
@@ -8,6 +8,8 @@
 /* Bert Truyens
 /************************************************************/
 
+require_once("classes/generalclasses.php");
+
 require_once("functions/olatBackupToOlatObject.php");
 require_once("functions/olatObjectToMoodleObject.php");
 require_once("functions/moodleObjectToMoodleBackup.php");
@@ -45,17 +47,20 @@ else {
 	$chapterFormat = "";
 }
 
+// The error handler initialization, this will contain all the errors at the end.
+$error = new o2mErrorHandler();
+
 if (isset($_FILES["file"])) {
-	if (file_exists($_FILES["file"]["tmp_name"]) || is_uploaded_file($_FILES["file"]["tmp_name"])) {
+	if (file_exists($_FILES["file"]["tmp_name"]) && is_uploaded_file($_FILES["file"]["tmp_name"])) {
 		// Creates an OLAT Object out of an exported OLAT course.
-		$olatObject = olatBackupToOlatObject($_FILES["file"]["tmp_name"]);
+		$olatObject = olatBackupToOlatObject($_FILES["file"]["tmp_name"], $error);
 		if ($olatObject !== null) {
 			echo "<br><p>===OLAT OBJECT===</p>";
 			echo "<p>OK - OLAT Object created</p><br>";
 			
 			echo "<p>===MOODLE OBJECT===</p>";
 			// Converts the OLAT Object to a Moodle object.
-			$moodleObject = olatObjectToMoodleObject($olatObject, $books);
+			$moodleObject = olatObjectToMoodleObject($olatObject, $error);
 			echo "<p>OK - Moodle Object created</p>";
 
 			if ($books) {
@@ -68,18 +73,44 @@ if (isset($_FILES["file"])) {
 
 			echo "<br><p>===MOODLE BACKUP===</p>";
 			// Uses the Moodle Object to make a Moodle backup .mbz file.
-			$moodleBackup = moodleObjectToMoodleBackup($moodleObject, $olatObject, $books, $chapterFormat);
+			$moodleBackup = moodleObjectToMoodleBackup($moodleObject, $olatObject, $books, $chapterFormat, $error);
 			echo "<p>OK - Moodle backup .mbz created</p><br>";
 
-			echo "<a href='" . dirname($_SERVER['PHP_SELF']) . $moodleBackup . "' class='download'>Download</a>";
+			echo "<a href='" . dirname($_SERVER['PHP_SELF']) . $moodleBackup . "' class='download'>Download</a><br>";
 		}
 	}
 	else {
-		echo "<p style='color:red;'>ERROR - No file uploaded.</p>";
+		$error->setError(new Error("ERROR", 5, "No file uploaded.", 0));
 	}
 }
 else {
-	echo "<p style='color:red;'>ERROR - No file found, did you land on this page by accident?</p>";
+	$error->setError(new Error("ERROR", 5, "No file found (or the file is too big)", 0));
+}
+
+$errors = $error->getErrors();
+if (empty($errors)) {
+	echo "<p style='color:green;'>OK - No warnings or errors found in the process.</p><br>";
+}
+else {
+	echo "<p style='color:darkorange;'>There were some issues with the course.</p>";
+	foreach ($errors as $e) {
+		if ($e->getPartOf() == 1) {
+			if ($e->getType() == "WARNING") {
+				echo "<p style='color:darkorange;margin-left:15px;'>" . $e->getErrorText() . "</p>";
+			}
+			else {
+				echo "<p style='color:red;margin-left:15px;'>" . $e->getErrorText() . "</p>";
+			}
+		}
+		else {
+			if ($e->getType() == "WARNING") {
+				echo "<p style='color:darkorange;'>" . $e->getErrorText() . "</p>";
+			}
+			else {
+				echo "<p style='color:red;'>" . $e->getErrorText() . "</p>";
+			}
+		}
+	}
 }
 
 echo '<footer>
